0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -22,7 +22,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%u] %s\n", (unsigned int)h->len, h->str);
 		}
 		node++;
 		h = h->next;
